Move BookSeriesUtil-code to BookSeries interface
