+++ build.gradle
@@ -44,7 +44,7 @@ def nvaCommonsVersion = '1.7.6'
 gradleLint.rules = ['unused-dependency']
 
 dependencies {
-    implementation group: 'commons-validator', name: 'commons-validator', version: '1.6'
+    implementation group: 'commons-validator', name: 'commons-validator', version: '1.7'
 
     compileOnly group: 'com.github.bibsysdev', name: 'core', version: nvaCommonsVersion
     testImplementation group: 'com.github.bibsysdev', name: 'core', version: nvaCommonsVersion
